Fixed warnings (#184)
* Fixed warnings

* Fixed CI error

* ruby

* Renames

* ruby

* hsp3

* Fixed hard coding file path

* Support metallicRoughnessTexture

* Support occlusionTexture

* Improved debug mode

* Fixed ternimation error on HSP3

+++ include/LuminoCore/Base/List.hpp
@@ -349,7 +349,7 @@ int List<T>::size() const LN_NOEXCEPT
 template<typename T>
 int List<T>::capacity() const LN_NOEXCEPT
 {
-    return m_data.capacity();
+    return static_cast<int>(m_data.capacity());
 }
 
 template<typename T>
